:books: docs: refactor: modulariza fluxo de compra e pagamento
